+++ frontend-two/types/stop-analysis/index.ts
@@ -1,25 +1,3 @@
-// Re-export generated types used by the stop analysis feature
-export type {
-  BoundingBoxInputType as BoundingBox,
-  DayOfWeekFlagsInputType as DayOfWeekFlags,
-  StopAnalysisQueryVariables as StopAnalysisFilters,
-} from "../../src/generated/graphql";
-export { MatchType } from "../../src/generated/graphql";
-
-export type StopStatistics =
-  import("../../src/generated/graphql").StopAnalysisQuery["stopAnalysis"][number];
-
-export type AdminArea = NonNullable<
-  import("../../src/generated/graphql").GetAdminAreasQuery["adminAreas"]
->[number];
-
-export type Operator =
-  import("../../src/generated/graphql").OperatorListQuery["operators"][number];
-
-export type Line =
-  import("../../src/generated/graphql").OperatorLinesQuery["lines"][number];
-
-// UI-only types not present in the GraphQL schema
 export type StopTypeOption = "TimingPoints" | "AllStops";
 
 export type Direction = "Inbound" | "Outbound";
